+++ src/Middleware.php
@@ -46,12 +46,42 @@ class Middleware
     }
 
     /**
+     * Read Maxemail's deprecation notices from the HTTP Warning header
+     * to log as warning and trigger deprecation notice
+     *
+     * @param HandlerStack $stack
+     * @param LoggerInterface $logger
+     */
+    public static function addWarningLogging(HandlerStack $stack, LoggerInterface $logger)
+    {
+        $middleware = GuzzleMiddleware::mapResponse(function (ResponseInterface $response) use ($logger) {
+            if ($response->hasHeader('Warning')) {
+                foreach ($response->getHeader('Warning') as $message) {
+                    // Code, agent, message, [date]
+                    $parts = str_getcsv($message, ' ', '"', '\\');
+                    if ($parts[0] != 299) {
+                        continue;
+                    }
+                    if (stripos($parts[1], 'mxmapi/') !== 0) {
+                        continue;
+                    }
+                    $logger->warning($parts[2]);
+                    trigger_error($parts[2], E_USER_DEPRECATED);
+                }
+            }
+
+            return $response;
+        });
+        $stack->push($middleware, 'mxm-deprecated');
+    }
+
+    /**
      * Add parser for Maxemail 4xx-level errors
      * @param HandlerStack $stack
      */
     public static function addMaxemailErrorParser(HandlerStack $stack)
     {
-        $middleware = GuzzleMiddleware::mapResponse(function (ResponseInterface $response) use ($stack) {
+        $middleware = GuzzleMiddleware::mapResponse(function (ResponseInterface $response) {
             $code = $response->getStatusCode();
             if ($code < 400 || $code >= 500) {
                 // Allow success response to continue, and 500-level errors to be handled by Guzzle