+++ tests/MiddlewareTest.php
@@ -11,6 +11,7 @@ use GuzzleHttp\HandlerStack;
 use GuzzleHttp\Psr7\Response;
 use Emailcenter\MaxemailApi\Exception\ClientException as MxmClientException;
 use PHPUnit\Framework\TestCase;
+use Psr\Log\LoggerInterface;
 
 class MiddlwareTest extends TestCase
 {
@@ -24,20 +25,121 @@ class MiddlwareTest extends TestCase
      */
     private $mockHandler;
 
+    /**
+     * @var HandlerStack
+     */
+    private $handlerStack;
+
     protected function setUp()
     {
         $this->mockHandler = new MockHandler();
 
-        $stack = HandlerStack::create($this->mockHandler);
-        Middleware::addMaxemailErrorParser($stack);
+        $this->handlerStack = HandlerStack::create($this->mockHandler);
 
         $this->httpClient = new GuzzleClient([
-            'handler' => $stack
+            'handler' => $this->handlerStack
         ]);
+    }
+
+    public function testWarningLoggerCreatesLog()
+    {
+        $warningMsg = 'dummyMethod Deprecated: some example description';
+        $warning = "299 MxmApi/v100 \"{$warningMsg}\"";
+
+        $this->expectException(\PHPUnit\Framework\Error\Deprecated::class);
+        $this->expectExceptionMessage($warningMsg);
+
+        /** @var LoggerInterface|\PHPUnit_Framework_MockObject_MockObject $logger */
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger->expects($this->once())
+            ->method('warning')
+            ->with($warningMsg);
+
+        Middleware::addWarningLogging($this->handlerStack, $logger);
+        $this->mockHandler->append(
+            new Response(200, ['Warning' => $warning], json_encode('OK'))
+        );
+
+        $service = new Service('dummy_service', $this->httpClient);
+        $service->dummyMethod();
+    }
+
+    public function testWarningLoggerSkipsNoAgent()
+    {
+        $warningMsg = 'some warning with no agent';
+        $warning = "299 - \"{$warningMsg}\"";
+
+        /** @var LoggerInterface|\PHPUnit_Framework_MockObject_MockObject $logger */
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger->expects($this->never())
+            ->method($this->anything());
+
+        Middleware::addWarningLogging($this->handlerStack, $logger);
+        $this->mockHandler->append(
+            new Response(200, ['Warning' => $warning], json_encode('OK'))
+        );
+
+        $service = new Service('dummy_service', $this->httpClient);
+        $service->dummyMethod();
+    }
+
+    public function testWarningLoggerSkipsWrongAgent()
+    {
+        $warningMsg = 'some other system';
+        $warning = "299 other/1.2.3 \"{$warningMsg}\"";
+
+        /** @var LoggerInterface|\PHPUnit_Framework_MockObject_MockObject $logger */
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger->expects($this->never())
+            ->method($this->anything());
+
+        Middleware::addWarningLogging($this->handlerStack, $logger);
+        $this->mockHandler->append(
+            new Response(200, ['Warning' => $warning], json_encode('OK'))
+        );
+
+        $service = new Service('dummy_service', $this->httpClient);
+        $service->dummyMethod();
+    }
+
+    public function testWarningLoggerSkipsWrongCode()
+    {
+        $warningMsg = 'something which looks like Maxemail';
+        $warning = "199 MxmApi/v100 \"{$warningMsg}\"";
+
+        /** @var LoggerInterface|\PHPUnit_Framework_MockObject_MockObject $logger */
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger->expects($this->never())
+            ->method($this->anything());
+
+        Middleware::addWarningLogging($this->handlerStack, $logger);
+        $this->mockHandler->append(
+            new Response(200, ['Warning' => $warning], json_encode('OK'))
+        );
+
+        $service = new Service('dummy_service', $this->httpClient);
+        $service->dummyMethod();
+    }
+
+    public function testWarningLoggerSkipsNoWarning()
+    {
+        /** @var LoggerInterface|\PHPUnit_Framework_MockObject_MockObject $logger */
+        $logger = $this->createMock(LoggerInterface::class);
+        $logger->expects($this->never())
+            ->method($this->anything());
+
+        Middleware::addWarningLogging($this->handlerStack, $logger);
+        $this->mockHandler->append(
+            new Response(200, [], json_encode('OK'))
+        );
+
+        $service = new Service('dummy_service', $this->httpClient);
+        $service->dummyMethod();
     }
 
     public function testErrorHandlerSkips200()
     {
+        Middleware::addMaxemailErrorParser($this->handlerStack);
         $this->mockHandler->append(
             new Response(200, [], json_encode('OK'))
         );
@@ -52,6 +154,7 @@ class MiddlwareTest extends TestCase
     {
         $this->expectException(ServerException::class);
 
+        Middleware::addMaxemailErrorParser($this->handlerStack);
         $this->mockHandler->append(
             new Response(500, [], 'Internal server error')
         );
@@ -64,6 +167,7 @@ class MiddlwareTest extends TestCase
     {
         $this->expectException(ClientException::class);
 
+        Middleware::addMaxemailErrorParser($this->handlerStack);
         $this->mockHandler->append(
             new Response(400, [], 'Error')
         );
@@ -79,6 +183,7 @@ class MiddlwareTest extends TestCase
         $this->expectException(MxmClientException::class);
         $this->expectExceptionMessage($errorMsg);
 
+        Middleware::addMaxemailErrorParser($this->handlerStack);
         $mxmError = [
             'success' => 'false',
             'msg' => $errorMsg